std: Move ptr to prim
